Moved `get_data.py` to `inversion_data` submodule
