--- openghg_inversions/data/__init__.py
@@ -1,3 +1,0 @@
-#!/usr/bin/env python
-
-from .serialise import load_merged_data
