+++ openghg_inversions/inversion_data/scenario.py
